+++ Modules/Core/ViewConfig.php
@@ -15,6 +15,7 @@ namespace D3\GoogleAnalytics4\Modules\Core;
 use D3\GoogleAnalytics4\Application\Model\ManagerHandler;
 use D3\GoogleAnalytics4\Application\Model\ManagerTypes;
 use OxidEsales\Eshop\Application\Controller\FrontendController;
+use OxidEsales\Eshop\Application\Model\User;
 use OxidEsales\Eshop\Core\Config;
 use OxidEsales\Eshop\Core\Registry;
 use OxidEsales\EshopCommunity\Internal\Container\ContainerFactory;
@@ -180,7 +181,8 @@ class ViewConfig extends ViewConfig_parent
 
         $oConfig = Registry::getConfig();
         $oView   = $oConfig->getTopActiveView();
-        /** @var FrontendController $oShop */
+
+        /** @var User|false|null $oUser */
         $oUser = $oConfig->getUser();
 
         $cl         = $this->getTopActiveClassName();
@@ -203,8 +205,8 @@ class ViewConfig extends ViewConfig_parent
                 'title' => $oView->getTitle(),
                 'cl'    => $cl,
             ],
-            'userid'    => ($oUser ? $oUser->getId() : false),
-            'sessionid' => session_id() ?? false,
+            'userid'    => $oUser instanceof User ? $oUser->getId() : false,
+            'sessionid' => session_id(),
             //'httpref'   => $_SERVER["HTTP_REFERER"] ?? "unknown"
         ];