+++ CHANGELOG.md
@@ -129,6 +129,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.0](https://git.d3data.de/D3Public/GoogleAnalytics4/compare/1.1...2.0) - 2023-01-20
 ### Added
 - using of ContainerFactory in ViewConfig
+## [1.15.0](https://git.d3data.de/D3Public/GoogleAnalytics4/compare/1.14.0...1.15.0) - 2024-01-25
+### Fixed
+- missing Component-ArticleDetails extension, missing ManufacturerListController extension
+
+## [1.14.0](https://git.d3data.de/D3Public/GoogleAnalytics4/compare/1.13.0...1.14.0) - 2024-01-03
+### Added
+- add_to_cart Event to several new shop-pages (landing-page, Account, category article-list, start)
+- PHPStan code-cleanup
+### Changed
+- template structure
+- code cleanup
+
 ## [1.13.0](https://git.d3data.de/D3Public/GoogleAnalytics4/compare/1.12.0...1.13.0) - 2023-09-13
 ### Added
 - extended Docs